Refactoring proposed by Jose Manuel Duarte to separate this module from the old implementation in org.biojava.nbio.genome.query.BlastXMLQuery.
