+++ biojava-core/src/main/resources/META-INF/services/org.biojava.nbio.core.search.io.ResultFactory
@@ -1,2 +1,2 @@
-org.biojava.nbio.core.search.io.blast.BlastXMLQuery
+org.biojava.nbio.core.search.io.blast.BlastXMLParser
 org.biojava.nbio.core.search.io.blast.BlastTabularParser
